Implements "improved" CMC logic on a list of CCF results for a comparison between two cartridge case scans as proposed by Tong et al. (2015)

Implements "improved' Congruent Matching Cells logic, as proposed by Tong et al. (2015), to the CCF results of a comparison between two cartridge case scans.

cmcFilter_improved(
  cellCCF_bothDirections_output,
  consensus_function = median,
  corr_thresh = 0.4,
  dx_thresh = 20,
  dy_thresh = dx_thresh,
  theta_thresh = 3,
  consensus_function_theta = consensus_function,
  ...
)

Arguments

cellCCF_bothDirections_output
list returned by the function cmcR::cellCCF_bothdirections

consensus_function
function to aggregate the translation (dx and dy) and rotation (theta) values in the ccfDF data frame to determine "consensus" values

corr_thresh
minimum correlation threshold to call a cell pair "congruent matching"

dx_thresh
maximum distance from the consensus dx value that a cell pair can be to be called "congruent matching"

dy_thresh
maximum distance from the consensus dy value that a cell pair can be to be called "congruent matching"

theta_thresh
maximum distance from the consensus theta value that a cell pair can be to be called "congruent matching"

consensus_function_theta
(OPTIONAL) function (separate from consensus_function) to aggregate the rotation (theta) values in the ccfDF data frame to determine "consensus" values

...
arguments to be passed to consensus_function or consensus_function_theta (e.g., na.rm = TRUE) if necessary

See also

Examples

if (FALSE) {
comparison1 <- cellCCF_bothDirections(x3p1,x3p2)

cmcs <- cmcFilter_improved(comparison1,
                           consensus_function = median,
                           corr_thresh = .4,
                           dx_thresh = 20,
                           dy_thresh = dx_thresh,
                           theta_thresh = 3,
                           consensus_function_theta = getMode)
}