Create a bar plot of congruent matching cells per rotation value

Create a bar plot of congruent matching cells per rotation value

Arguments

cellCCF_output
list returned by the cellCCF or cellCCF_bothDirections function. If from the cellCCF_bothDirections, then the ggplot will be faceted by the "direction" of the comparison (i.e., whether x3p1 or x3p2 played the role as the "questioned" cartridge case scan)

consensus_function
function to aggregate the translation (dx and dy) and rotation (theta) values in the ccfDF data frame to determine "consensus" values

ccf_thresh
minimum correlation threshold to call a cell pair "congruent matching"

dx_thresh
maximum distance from the consensus dx value that a cell pair can be to be called "congruent matching"

dy_thresh
maximum distance from the consensus dy value that a cell pair can be to be called "congruent matching"

theta_thresh
maximum distance from the consensus theta value that a cell pair can be to be called "congruent matching"

consensus_function_theta
(OPTIONAL) function (separate from consensus_function) to aggregate the rotation (theta) values in the ccfDF data frame to determine "consensus" values

highCMCThresh
number of CMCs less than the max CMC count that should be classified as a "High CMC" count. That is, CMC_high = CMC_max - highCMCThresh

x3pNames
(Optional) Names of x3p objects to be included in x3pListPlot function

Examples

if (FALSE) {
 #x3p1 and x3p2 are two x3p objects containing processed cartridge case scans
 comparison1 <- cellCCF(x3p1,x3p2)

 #creates a single bar plot of the CMCs per theta
 cmcPerThetaBarPlot(comparison1)

 comparison2 <- cellCCF_bothDirections(x3p1,x3p2)

 #creates a faceted bar plot of the CMCs per theta in both comparison directions
 cmcPerThetaBarPlot(comparison2)
}