Performs all steps in the cell-based comparison procedure.

Performs all steps in the cell-based comparison procedure.

comparison_allTogether(
  reference,
  target,
  theta = 0,
  numCells = 64,
  maxMissingProp = 0.85
)

Arguments

reference
an x3p object containing a breech face scan to be treated as the "reference scan" partitioned into a grid of cells

target
an x3p object containing a breech face scan to be treated as the "target scan" that the reference scan's cells are compared to

theta
degrees that the target scan is to be rotated prior extracting regions.

numCells
number of cells to partition the breech face scan into. Must be a perfect square (49, 64, 81, etc.)

maxMissingProp
maximum proportion of missing values allowed for each cell/region.

data(fadul1.1_processed,fadul1.2_processed)

comparisonDF <- comparison_allTogether(reference = fadul1.1_processed, target = fadul1.2_processed)

head(comparisonDF)

Value

a tibble object containing cell indices and the x, y, FFT-based CCF, and pairwise-complete correlation associated with the comparison between each cell and its associated target scan region (after rotating the target scan by theta degrees)

Examples

data(fadul1.1_processed,fadul1.2_processed)

cellTibble <- comparison_allTogether(reference = fadul1.1_processed,target = fadul1.2_processed)

head(cellTibble)

#> # A tibble: 6 x 6
#>   cellIndex     x     y fft_ccf pairwiseCompCor theta
#>   <chr>     <dbl> <dbl>   <dbl>           <dbl> <dbl>
#> 1 8, 6         -7   -23   0.241           0.509     0
#> 2 7, 7         55    31   0.186           0.331     0
#> 3 7, 8         53    53   0.191           0.490     0
#> 4 6, 8        -23   -60   0.166           0.345     0
#> 5 5, 8         -3    13   0.175           0.368     0
#> 6 4, 1        -14   -48   0.209           0.417     0