Calculates correlation between a cell and a matrix of the same dimensions extracted from the cell's associated region.

Calculates correlation between a cell and a matrix of the same dimensions extracted from the cell's associated region.

comparison_cor(
  cellHeightValues,
  regionHeightValues,
  fft_ccf_df,
  use = "pairwise.complete.obs"
)

Arguments

cellHeightValues

list/tibble column of x3p objects containing a reference scan's cells (as returned by comparison_cellDivision)

regionHeightValues

list/tibble column of x3p objects containing a target scan's regions (as returned by comparison_getTargetRegions)

fft_ccf_df

data frame/tibble column containing the data frame of (x,y) and CCF values returned by comparison_fft_ccf

use

argument for stats::cor

Value

A vector of the same length as the input containing correlation values at the estimated alignment between each reference cell and its associated target region

Examples

data(fadul1.1_processed,fadul1.2_processed)

cellTibble <- fadul1.1_processed %>%
comparison_cellDivision(numCells = 64) %>%
dplyr::mutate(regionHeightValues =
             comparison_getTargetRegions(cellHeightValues = cellHeightValues,
                                         target = fadul1.2_processed)) %>%
dplyr::mutate(cellPropMissing =
           comparison_calcPropMissing(heightValues = cellHeightValues),
              regionPropMissing =
           comparison_calcPropMissing(heightValues = regionHeightValues)) %>%
dplyr::filter(cellPropMissing <= .85 & regionPropMissing <= .85) %>%
dplyr::mutate(cellHeightValues =
        comparison_standardizeHeights(heightValues = cellHeightValues),
              regionHeightValues =
        comparison_standardizeHeights(heightValues = regionHeightValues)) %>%
dplyr::mutate(cellHeightValues =
                  comparison_replaceMissing(heightValues = cellHeightValues),
              regionHeightValues =
            comparison_replaceMissing(heightValues = regionHeightValues)) %>%
dplyr::mutate(fft_ccf_df = comparison_fft_ccf(cellHeightValues,
                                              regionHeightValues)) %>%
dplyr::mutate(pairwiseCompCor = comparison_cor(cellHeightValues,
                                               regionHeightValues,
                                               fft_ccf_df))

head(cellTibble)
#> # A tibble: 6 x 7
#>   cellIndex cellHeightValues regionHeightValues cellPropMissing regionPropMissi~
#>   <chr>     <named list>     <named list>                 <dbl>            <dbl>
#> 1 8, 6      <x3p>            <x3p>                        0.833            0.808
#> 2 7, 7      <x3p>            <x3p>                        0.657            0.700
#> 3 7, 8      <x3p>            <x3p>                        0.834            0.757
#> 4 6, 8      <x3p>            <x3p>                        0.353            0.638
#> 5 5, 8      <x3p>            <x3p>                        0.153            0.576
#> 6 4, 1      <x3p>            <x3p>                        0.117            0.767
#> # ... with 2 more variables: fft_ccf_df <named list>, pairwiseCompCor <dbl>