Identifies the congruent matching cells between two processed cartridge case scans

Wrapper for calling the cellCCF_bothDirections and cmcFilter_improved functions together

congruentMatchingCells(
  x3p1,
  x3p2,
  thetas = seq(-30, 30, by = 3),
  cellNumHoriz = 8,
  cellNumVert = cellNumHoriz,
  regionToCellProp = 4,
  minObservedProp = 0.15,
  centerCell = "individualCell",
  scaleCell = "individualCell",
  consensus_function = median,
  ccf_thresh = 0.6,
  dx_thresh = 10,
  dy_thresh = dx_thresh,
  theta_thresh = 3,
  consensus_function_theta = consensus_function,
  ...
)

Arguments

x3p1
(no default) an x3p object containing the surface matrix of a breech face impression

x3p2
(no default) an x3p object containing the surface matrix of a breech face impression to be compared to that in x3p1

thetas
(default seq(from = -30,to = 30,by = 3)) rotation values (in degrees) for which x3p2$surface.matrix will be rotated, split into cells, and compared to x3p1$surface.matrix

cellNumHoriz
(default 7) number of cells along horizontal axis to divide x3p1$surface.matrix into

cellNumVert
(default equal to cellNumHoriz) number of cells along vertical axis to divide x3p1$surface.matrix into

regionToCellProp
determines how much larger the x3p2 regions will be relative to the x3p1 cells. For example, if regionToCellProp = 4 means that the x3p2 regions will be 4 times times larger (sidelengths multiplied by 2)

minObservedProp
(default 15) the minimum proportion of a cell that needs to contain observed (i.e., non-NA) values for it to be included in the CCF calculation procedure

centerCell
dictates if cell is to be shifted prior to CCF calculation. Default is that no shifting is performed. If set to "wholeMatrix", then each cell is subracted by the mean of the whole matrix. If set to "individualCell", then each cell is subtracted by its particular mean.

scaleCell
dictates if cell is to be scaled prior to CCF calculation. Default is that no scaling is performed. If set to "wholeMatrix", then each cell is divided by the standard deviation of the whole matrix. If set to "individualCell", then each cell is divided by its particular standard deviation.

consensus_function
function to aggregate the translation (dx and dy) and rotation (theta) values in the ccfDF data frame to determine "consensus" values

ccf_thresh
minimum correlation threshold to call a cell pair "congruent matching"

dx_thresh
maximum distance from the consensus dx value that a cell pair can be to be called "congruent matching"

dy_thresh
maximum distance from the consensus dy value that a cell pair can be to be called "congruent matching"

theta_thresh
maximum distance from the consensus theta value that a cell pair can be to be called "congruent matching"

consensus_function_theta
(OPTIONAL) function (separate from consensus_function) to aggregate the rotation (theta) values in the ccfDF data frame to determine "consensus" values

...
arguments to be passed to consensus_function or consensus_function_theta (e.g., na.rm = TRUE) if necessary

See also

Examples

if (FALSE) {
#x3p1 and x3p2 are two processed cartridge case scans
cmc <- congruentMatchingCells(x3p1,x3p2)
}