Combine CMC data from two comparison directions

Combine CMC data from two comparison directions

decision_combineCMCDirections(
  comparison_1to2_df,
  comparison_2to1_df,
  corColName = "pairwiseCompCor",
  method = "highCMC",
  compareThetas = TRUE,
  thetaThresh = 6
)

Arguments

comparison_1to2_df
a data frame containing a column "originalMethodClassif" or "highCMCClassif" (or both) representing CMCs identified for a comparison between two cartridge cases in which one is considered the "reference scan" and the other the "target scan". The decision_CMC function can be used to add CMC column(s) to a data frame.

comparison_2to1_df
a data frame containing a column "originalMethodClassif" or "highCMCClassif" (or both) representing CMCs identified for a comparison between two cartridge cases where the roles of "reference" and "target" are reversed relative to the comparison_1to2_df CMCs

corColName
name of correlation similarity score column used to identify the CMCs in the two comparison_*_df data frames (e.g., pairwiseCompCor)

method
whether to apply the combination logic of the High CMC method or the original method of Song (2013)

compareThetas
controls whether the estimated "true" rotation values calculated under the original method of Song (2013) and the High CMC method are compared to each other (they should be similar for a truly matching scan).

thetaThresh
threshold to use when comparing the estimated theta values under the original method of Song (2013) and the High CMC method. Also used to compare the estimated theta values in both comparison directions under the High CMC method (as they should be approximate opposites of each other).

Examples

data(fadul1.1_processed,fadul1.2_processed)

comparisonDF_1to2 <- purrr::map_dfr(seq(-30,30,by = 3),
                                   ~ comparison_allTogether(fadul1.1_processed,
                                                       fadul1.2_processed,
                                                       theta = .))
comparisonDF_2to1 <- purrr::map_dfr(seq(-30,30,by = 3),
                                   ~ comparison_allTogether(fadul1.2_processed,
                                                       fadul1.1_processed,
                                                       theta = .))

comparisonDF_1to2 <- comparisonDF_1to2 %>%
dplyr::mutate(originalMethodClassif = decision_CMC(cellIndex = cellIndex,
                                                   x = x,
                                                   y = y,
                                                   theta = theta,
                                                   corr = pairwiseCompCor),
              highCMCClassif = decision_CMC(cellIndex = cellIndex,
                                           x = x,
                                           y = y,
                                           theta = theta,
                                           corr = pairwiseCompCor,
                                           tau = 1))

comparisonDF_2to1 <- comparisonDF_2to1 %>%
dplyr::mutate(originalMethodClassif = decision_CMC(cellIndex = cellIndex,
                                                   x = x,
                                                   y = y,
                                                   theta = theta,
                                                   corr = pairwiseCompCor),
              highCMCClassif = decision_CMC(cellIndex = cellIndex,
                                           x = x,
                                           y = y,
                                           theta = theta,
                                           corr = pairwiseCompCor,
                                           tau = 1))

decision_combineCMCDirections(comparisonDF_1to2,
                             comparisonDF_2to1,
                             corColName = "pairwiseCompCor",
                             method = "highCMC")

#> Error in bind_rows(comparison_1to2_df %>% dplyr::filter(originalMethodClassif ==     "CMC" | highCMCClassif == "CMC") %>% dplyr::mutate(direction = "comparison_AtoB"),     comparison_2to1_df %>% dplyr::filter(originalMethodClassif ==         "CMC" | highCMCClassif == "CMC") %>% dplyr::mutate(direction = "comparison_BtoA")): could not find function "bind_rows"