Apply CMC classification logic of the Tong et al. (2015) to the CMC-theta distribution returned by the decision_highCMC_cmcThetaDistrib function

Apply CMC classification logic of the Tong et al. (2015) to the CMC-theta distribution returned by the decision_highCMC_cmcThetaDistrib function

decision_highCMC_classifyCMCs(
  cellIndex,
  x,
  y,
  theta,
  corr,
  xThresh = 20,
  yThresh = xThresh,
  thetaThresh = 6,
  corrThresh = 0.5,
  tau = 1
)

Arguments

cellIndex
vector/tibble column containing cell indices corresponding to a reference cell

x
vector/tibble column containing x horizontal translation values

y
vector/tibble column containing y vertical translation values

theta
vector/tibble column containing theta rotation values

corr
vector/tibble column containing correlation similarity scores between a reference cell and its associated target region

xThresh
used to classify particular x values "congruent" (conditional on a particular theta value) if they are within xThresh of the theta-specific median x value

yThresh
used to classify particular y values "congruent" (conditional on a particular theta value) if they are within yThresh of the theta-specific median y value

corrThresh
to classify particular correlation values "congruent" (conditional on a particular theta value) if they are at least corrThresh

tau
constant used to define a "high" CMC count. This number is subtracted from the maximum CMC count achieved in the CMC-theta distribution. Theta values with CMC counts above this value are considered to have "high" CMC counts.

Note

the decision_CMC function internally calls this function if a value of tau is provided

See also

Examples

data(fadul1.1_processed,fadul1.2_processed)

comparisonDF <- purrr::map_dfr(seq(-30,30,by = 3),
                              ~ comparison_allTogether(fadul1.1_processed,
                                                       fadul1.2_processed,
                                                       theta = .))

comparisonDF <- comparisonDF %>%
dplyr::mutate(highCMCClassif = decision_highCMC_classifyCMCs(cellIndex = cellIndex,
                                                             x = x,
                                                             y = y,
                                                             theta = theta,
                                                             corr = pairwiseCompCor))

comparisonDF %>%
dplyr::filter(highCMCClassif == "CMC")

#> # A tibble: 19 x 7
#>    cellIndex     x     y fft_ccf pairwiseCompCor theta highCMCClassif
#>    <chr>     <dbl> <dbl>   <dbl>           <dbl> <dbl> <chr>         
#>  1 6, 1        -13    26   0.436           0.760   -27 CMC           
#>  2 5, 1         -6    23   0.333           0.720   -27 CMC           
#>  3 5, 8         -8   -12   0.244           0.602   -27 CMC           
#>  4 3, 1          3    22   0.439           0.782   -27 CMC           
#>  5 3, 8          1    -6   0.333           0.600   -27 CMC           
#>  6 2, 2          5    18   0.252           0.697   -27 CMC           
#>  7 2, 5          7     7   0.187           0.503   -27 CMC           
#>  8 2, 7          6    -7   0.351           0.721   -27 CMC           
#>  9 6, 8         -8     4   0.273           0.609   -24 CMC           
#> 10 4, 1         -2    11   0.366           0.699   -24 CMC           
#> 11 4, 8         -5     1   0.255           0.616   -24 CMC           
#> 12 3, 2         -2     9   0.294           0.793   -24 CMC           
#> 13 2, 3         -1     6   0.195           0.664   -24 CMC           
#> 14 2, 4         -3     8   0.200           0.595   -24 CMC           
#> 15 1, 3         -1    11   0.295           0.673   -24 CMC           
#> 16 1, 4         -2     7   0.255           0.634   -24 CMC           
#> 17 1, 5         -2     7   0.248           0.525   -24 CMC           
#> 18 1, 6         -1     7   0.294           0.658   -24 CMC           
#> 19 1, 7         -1     0   0.362           0.880   -24 CMC