Crop the exterior of a breech face impression surface matrix

The radius estimation procedure tends to over-estimate the desired radius values. As such, a lot of the breech face impression "roll-off" is included in the final scan. Excessive roll-off can bias the calculation of the CCF. As such, we can manually shrink the radius estimate so that little to no roll-off is included in the final processed scan.

preProcess_cropExterior(
  x3p,
  scheme = 3,
  high_connectivity = FALSE,
  tolerance = 0,
  radiusOffset = 0,
  croppingThresh = 1,
  agg_function = median
)

Arguments

x3p
an x3p object containing the surface matrix of a cartridge case scan

scheme
argument for imager::imgradient

high_connectivity
argument for imager::label

tolerance
argument for imager::label

radiusOffset
number of pixels to add to estimated breech face radius. This is commonly a negative value (e.g., -20) to trim the cartridge case primer roll-off from the returned, cropped surface matrix.

croppingThresh
argument for cmcR::preProcess_cropWS

agg_function
the breech face radius estimation procedure returns numerous radius estimates. This argument dictates the function used to aggregate these into a final estimate.

Value

An x3p object containing the surface matrix of a breech face impression scan where the rows/columns on the exterior of the breech face impression have been cropped-out.