Finds plane of breechface marks using the RANSAC method

Given input depths (in microns), find best-fitting plane using RANSAC. This should be the plane that the breechface marks are on. Adapted from cartridges3D::findPlaneRansac function.

preProcess_ransac(
  surfaceMat,
  inlierTreshold = (10^(-5)),
  finalSelectionThreshold = 2 * (10^(-5)),
  iters = 150
)

Arguments

surfaceMat
a surface matrix representing a breech face impression scan

finalSelectionThreshold
once the RANSAC plane is fitted based on the inlierThreshold, this argument dictates which observations are selected as the final breech face estimate.

iters
number of candidate planes to fit (higher value yields more stable breech face estimate)

inlierThreshold
threshold to declare an observed value close to the fitted plane an "inlier". A smaller value will yield a more stable estimate.

Value

List object containing fitted plane (as an lm object) and selected breechface marks (matrix of same size as original matrix containing all inliers close to fitted plane).

See also

Examples

if (FALSE) {
raw_x3p <- x3ptools::read_x3p("path/to/file.x3p") %>%
  x3ptools::sample_x3p(m = 2)

fittedPlane <- raw_x3p$surface.matrix %>%
  preProcess_ransac(inlierTreshold = 10^(-5),
                    finalSelectionThreshold = 2*10^(-5),
                    iters = 150)
}