Select breech face impression from a cartridge case scan after using imager::resize to resize the scan

Select breech face impression from a cartridge case scan after using imager::resize to resize the scan

selectBFImpression_resize(
  x3p_path,
  ransacInlierThresh = (10^(-5)),
  ransacFinalSelectThresh = 2 * (10^(-5)),
  ransacIters = 150,
  useResiduals = FALSE,
  croppingThresh = 1,
  standardizeBF = FALSE,
  size_x,
  size_y = size_x,
  interpolation_type = 1,
  boundary_conditions = 0
)

Arguments

x3p_path
path to a .x3p file

ransacInlierThresh
threshold to declare an observed value close to the fitted plane an "inlier" for the RANSAC method

ransacIters
number of candidate planes to fit for the RANSAC method (higher value yields more stable breech face estimate)

useResiduals
dictates whether the difference between the estimated breech face and fitted plane are returned (residuals) or if the estimates breech face is simply shifted down by its mean value

croppingThresh
treshold for cropping exterior whitespace from the breech face impression

size_x
Number of columns (new size along the X-axis). Note: the "X"-axis according to imager is actually the rows of the matrix

size_y
Number of rows (new size along the Y-axis). Note: the "Y"-axis according to imager is actually the number of cols of the matrix

interpolation_type
Method of interpolation: -1 = no interpolation: raw memory resizing. 0 = no interpolation: additional space is filled according to boundary_conditions. 1 = nearest-neighbor interpolation. 2 = moving average interpolation. 3 = linear interpolation. 4 = grid interpolation. 5 = cubic interpolation. 6 = lanczos interpolation.

boundary_conditions
Border condition type.

standardizedBF
subtract mean and divide by standard deviation of processed surface matrix

Value

Note

imager treats a matrix as its transpose (i.e., x and y axes are swapped). As such the size_x argument corresponds to changing the number of rows in the original matrix and size_y the number of columns.

See also

Examples

if (FALSE) {

}