Implements "improved" CMC logic on a list of CCF results for a comparison between two cartridge case scans as proposed by Tong et al. (2015)

Implements "improved' Congruent Matching Cells logic, as proposed by Tong et al. (2015), to the CCF results of a comparison between two cartridge case scans.

Arguments

cellCCF_bothDirections_output
list returned by the function cmcR::cellCCF_bothdirections

consensus_function
function to aggregate the translation (dx and dy) and rotation (theta) values in the ccfDF data frame to determine "consensus" values

ccf_thresh
minimum correlation threshold to call a cell pair "congruent matching"

dx_thresh
maximum distance from the consensus dx value that a cell pair can be to be called "congruent matching"

dy_thresh
maximum distance from the consensus dy value that a cell pair can be to be called "congruent matching"

theta_thresh
maximum distance from the consensus theta value that a cell pair can be to be called "congruent matching"

missingTheta_decision
dictates how function should handle situations in which one direction passes the high CMC criterion while another direction does not. "replace": replaces theta value in failed direction with opposite of theta value in successful direction. "dismiss": only counts the initial CMCs in failed direction and high CMCs in successful direction. "fail": only counts the initial CMCs in either direction.

consensus_function_theta
(OPTIONAL) function (separate from consensus_function) to aggregate the rotation (theta) values in the ccfDF data frame to determine "consensus" values

See also

Examples

if (FALSE) {
comparison1 <- cellCCF_bothDirections(x3p1,x3p2)

cmcs <- cmcFilter_improved(comparison1,
                           consensus_function = median,
                           ccf_thresh = .4,
                           dx_thresh = 20,
                           dy_thresh = dx_thresh,
                           theta_thresh = 3,
                           missingTheta_decision = "replace",
                           consensus_function_theta = getMode)
}